Remove locale dependence in parsed_apache_config test
This corrects the problem raised in GitHub issue #5.  By not relying on the
locale-based error output in the test, one can then still check that the
test dies as expected, and with the expected error message, however still
allow the tests to run correctly on systems with a non-English locale.